requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Django==2.2.17
idna==2.10
joblib==0.17.0
lxml==4.6.1
mysql-connector-python==8.0.22
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
pandas==1.1.4
protobuf==3.14.0
PyMySQL==0.10.1
PySastrawi==1.2.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
Sastrawi==1.0.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.53.0
tweepy==3.9.0
tweet-preprocessor==0.6.0
urllib3==1.26.2
wheel==0.36.1